requirements.txt
beautifulsoup4==4.7.1
numpy==1.16.3
pandas==0.24.2
feedparser==5.2.1
pygame==1.9.6
python-dateutil==2.8.0
requests==2.21.0
